+++ R/zzz.R
@@ -6,19 +6,6 @@
 
   run_on_load()
 
-  on_package_load("testthat", {
-    s3_register(
-      "testthat::is_informative_error",
-      "vctrs_error_cast_lossy",
-      is_informative_error_vctrs_error_cast_lossy
-    )
-    s3_register(
-      "testthat::is_informative_error",
-      "vctrs_error_cast_lossy_dropped",
-      is_informative_error_vctrs_error_cast_lossy_dropped
-    )
-  })
-
   s3_register("generics::as.factor", "vctrs_vctr")
   s3_register("generics::as.ordered", "vctrs_vctr")
   s3_register("generics::as.difftime", "vctrs_vctr")